requirements.txt
numpy==1.23.5
Pillow==9.5.0
streamlit==1.20.0
tensorflow==2.10.0
